Replace existing lithology table
