--- src/client/src/pages/detail/form/stratigraphy/lithology/lithologyContentEdit.tsx
@@ -1,583 +1,0 @@
-import { Dispatch, FC, ReactNode, SetStateAction, useCallback, useContext, useEffect, useMemo, useState } from "react";
-import { useTranslation } from "react-i18next";
-import { Stack, Typography } from "@mui/material";
-import { BaseLayer, MinimalLayer } from "../../../../../api/stratigraphy.ts";
-import { AlertContext } from "../../../../../components/alert/alertContext.tsx";
-import { AddRowButton } from "../../../../../components/buttons/buttons.tsx";
-import { formatNumberForDisplay } from "../../../../../components/form/formUtils.ts";
-import { SaveContext } from "../../../saveContext.tsx";
-import { FaciesDescription, useFaciesDescriptionMutations } from "../faciesDescription";
-import { LithologicalDescription, useLithologicalDescriptionMutations } from "../lithologicalDescription.ts";
-import { LayerDepth, Lithology, useLithologyMutations } from "../lithology.ts";
-import { StratigraphyContext, StratigraphyContextProps } from "../stratigraphyContext";
-import {
-  LayerAddButton,
-  StratigraphyTableActionCell,
-  StratigraphyTableCell,
-  StratigraphyTableColumn,
-  StratigraphyTableContent,
-  StratigraphyTableDescriptionGap,
-  StratigraphyTableGap,
-  StratigraphyTableHeader,
-  StratigraphyTableHeaderCell,
-} from "../stratigraphyTableComponents.tsx";
-import {
-  BaseLayerChangeTracker,
-  computeCellHeight,
-  defaultRowHeight,
-  getLayerDepths,
-  getLayersWithGaps,
-} from "../stratigraphyUtils.ts";
-import { FaciesDescriptionLabels } from "./faciesDescriptionLabels.tsx";
-import { FaciesDescriptionModal } from "./form/faciesDescriptionModal.tsx";
-import { LithologicalDescriptionModal } from "./form/lithologicalDescriptionModal.tsx";
-import { LithologyModal } from "./form/lithologyModal.tsx";
-import { LithologyLabels } from "./lithologyLabels.tsx";
-
-interface LithologyContentEditProps {
-  stratigraphyId: number;
-  lithologies: Lithology[];
-  lithologicalDescriptions: LithologicalDescription[];
-  faciesDescriptions: FaciesDescription[];
-}
-
-export const LithologyContentEdit: FC<LithologyContentEditProps> = ({
-  stratigraphyId,
-  lithologies,
-  lithologicalDescriptions,
-  faciesDescriptions,
-}) => {
-  const { t } = useTranslation();
-  const { showAlert } = useContext(AlertContext);
-  const { registerSaveHandler, registerResetHandler } = useContext<StratigraphyContextProps>(StratigraphyContext);
-  const { markAsChanged } = useContext(SaveContext);
-  const {
-    add: { mutateAsync: addLithology, isPending: isAddingLithology },
-    update: { mutateAsync: updateLithology, isPending: isUpdatingLithology },
-    delete: { mutateAsync: deleteLithology, isPending: isDeletingLithology },
-    invalidateQueries: invalidateLithologyQueries,
-  } = useLithologyMutations(true);
-  const {
-    add: { mutateAsync: addLithologicalDescription, isPending: isAddingLithologicalDescription },
-    update: { mutateAsync: updateLithologicalDescription, isPending: isUpdatingLithologicalDescription },
-    delete: { mutateAsync: deleteLithologicalDescription, isPending: isDeletingLithologicalDescription },
-    invalidateQueries: invalidateLithologicalDescriptionQueries,
-  } = useLithologicalDescriptionMutations(true);
-  const {
-    add: { mutateAsync: addFaciesDescription, isPending: isAddingFaciesDescription },
-    update: { mutateAsync: updateFaciesDescription, isPending: isUpdatingFaciesDescription },
-    delete: { mutateAsync: deleteFaciesDescription, isPending: isDeletingFaciesDescription },
-    invalidateQueries: invalidateFaciesDescriptionQueries,
-  } = useFaciesDescriptionMutations(true);
-
-  const [tmpLithologies, setTmpLithologies] = useState<BaseLayerChangeTracker[]>([]);
-  const tmpLithologiesFlat = useMemo(() => tmpLithologies.map(l => l.item as Lithology), [tmpLithologies]);
-  const [tmpLithologicalDescriptions, setTmpLithologicalDescriptions] = useState<BaseLayerChangeTracker[]>([]);
-  const tmpLithologicalDescriptionsFlat = useMemo(
-    () => tmpLithologicalDescriptions.map(l => l.item as LithologicalDescription),
-    [tmpLithologicalDescriptions],
-  );
-  const [tmpFaciesDescriptions, setTmpFaciesDescriptions] = useState<BaseLayerChangeTracker[]>([]);
-  const tmpFaciesDescriptionsFlat = useMemo(
-    () => tmpFaciesDescriptions.map(l => l.item as FaciesDescription),
-    [tmpFaciesDescriptions],
-  );
-
-  const [depths, setDepths] = useState<LayerDepth[]>([]);
-
-  const [selectedLithologyIndex, setSelectedLithologyIndex] = useState<number>();
-  const selectedLithology = useMemo(() => {
-    if (selectedLithologyIndex === undefined) return undefined;
-    if (selectedLithologyIndex === -1) {
-      return {
-        fromDepth: depths?.at(-1)?.toDepth,
-        toDepth: undefined,
-        id: 0,
-        isGap: true,
-        stratigraphyId: 0,
-        isUnconsolidated: tmpLithologiesFlat?.at(-1)?.isUnconsolidated ?? true,
-        hasBedding: false,
-      } as MinimalLayer as Lithology;
-    }
-    return tmpLithologiesFlat.at(selectedLithologyIndex);
-  }, [depths, selectedLithologyIndex, tmpLithologiesFlat]);
-
-  const [selectedLithologicalDescriptionIndex, setSelectedLithologicalDescriptionIndex] = useState<number>();
-  const selectedLithologicalDescription = useMemo(() => {
-    if (selectedLithologicalDescriptionIndex === undefined) return undefined;
-    return tmpLithologicalDescriptionsFlat.at(selectedLithologicalDescriptionIndex);
-  }, [selectedLithologicalDescriptionIndex, tmpLithologicalDescriptionsFlat]);
-
-  const [selectedFaciesDescriptionIndex, setSelectedFaciesDescriptionIndex] = useState<number>();
-  const selectedFaciesDescription = useMemo(() => {
-    if (selectedFaciesDescriptionIndex === undefined) return undefined;
-    return tmpFaciesDescriptionsFlat.at(selectedFaciesDescriptionIndex);
-  }, [selectedFaciesDescriptionIndex, tmpFaciesDescriptionsFlat]);
-
-  const updateStratigraphyItem = useCallback(
-    (
-      index: number | undefined,
-      item: BaseLayer,
-      setState: Dispatch<SetStateAction<BaseLayerChangeTracker[]>>,
-      hasChanges: boolean,
-    ) => {
-      setState(prev => {
-        if (hasChanges) {
-          markAsChanged(true);
-          item.isGap = false;
-
-          if (typeof index === "number" && index >= 0 && index < prev.length) {
-            const updated = [...prev];
-
-            updated[index] = { item, hasChanges: true };
-            return updated;
-          }
-
-          // fallback to add if index is undefined or out of bounds
-          return [...prev, { item, hasChanges: true }];
-        }
-
-        return prev;
-      });
-    },
-    [markAsChanged],
-  );
-
-  const updateTmpLithology = useCallback(
-    (lithology: Lithology, hasChanges: boolean) => {
-      updateStratigraphyItem(selectedLithologyIndex, lithology, setTmpLithologies, hasChanges);
-      setSelectedLithologyIndex(undefined);
-    },
-    [selectedLithologyIndex, updateStratigraphyItem],
-  );
-
-  const handleEditLithology = useCallback((layerIndex: number) => {
-    setSelectedLithologyIndex(layerIndex);
-  }, []);
-
-  const handleDeleteLithology = useCallback(
-    (index: number) => {
-      setTmpLithologies(prev => {
-        if (index >= 0 && index < prev.length) {
-          const newTmpLithologies = [...prev.slice(0, index), ...prev.slice(index + 1)];
-          markAsChanged(true);
-          return newTmpLithologies;
-        }
-
-        return prev;
-      });
-    },
-    [markAsChanged],
-  );
-
-  const updateTmpLithologicalDescription = useCallback(
-    (lithologicalDescription: LithologicalDescription, hasChanges: boolean) => {
-      updateStratigraphyItem(
-        selectedLithologicalDescriptionIndex,
-        lithologicalDescription,
-        setTmpLithologicalDescriptions,
-        hasChanges,
-      );
-      setSelectedLithologicalDescriptionIndex(undefined);
-    },
-    [selectedLithologicalDescriptionIndex, updateStratigraphyItem],
-  );
-
-  const handleEditLithologicalDescription = useCallback((index: number) => {
-    setSelectedLithologicalDescriptionIndex(index);
-  }, []);
-
-  const handleDeleteLithologicalDescription = useCallback(
-    (index: number) => {
-      setTmpLithologicalDescriptions(prev => {
-        if (index >= 0 && index < prev.length) {
-          const newTmpLithologicalDescriptions = [...prev.slice(0, index), ...prev.slice(index + 1)];
-          markAsChanged(true);
-          return newTmpLithologicalDescriptions;
-        }
-
-        return prev;
-      });
-    },
-    [markAsChanged],
-  );
-
-  const updateTmpFaciesDescription = useCallback(
-    (faciesDescription: FaciesDescription, hasChanges: boolean) => {
-      updateStratigraphyItem(selectedFaciesDescriptionIndex, faciesDescription, setTmpFaciesDescriptions, hasChanges);
-      setSelectedFaciesDescriptionIndex(undefined);
-    },
-    [selectedFaciesDescriptionIndex, updateStratigraphyItem],
-  );
-
-  const handleEditFaciesDescription = useCallback((index: number) => {
-    setSelectedFaciesDescriptionIndex(index);
-  }, []);
-
-  const handleDeleteFaciesDescription = useCallback(
-    (index: number) => {
-      setTmpFaciesDescriptions(prev => {
-        if (index >= 0 && index < prev.length) {
-          const newTmpFaciesDescriptions = [...prev.slice(0, index), ...prev.slice(index + 1)];
-          markAsChanged(true);
-          return newTmpFaciesDescriptions;
-        }
-
-        return prev;
-      });
-    },
-    [markAsChanged],
-  );
-
-  const getDepthOptions = useCallback(
-    (layer: BaseLayer | undefined, layers: BaseLayer[], getForToDepths: boolean = false) => {
-      if (!layer || !tmpLithologiesFlat) return [];
-      let allPossibleDepths: number[] = [tmpLithologiesFlat[0].fromDepth, ...tmpLithologiesFlat.map(c => c.toDepth)];
-      const layerIndex = layers.findIndex(l => l.fromDepth === layer.fromDepth && l.toDepth === layer.toDepth);
-      // Walk past any gap layers when restricting depth options (description columns can have multiple consecutive gaps)
-      let previousIndex = layerIndex - 1;
-      while (layers[previousIndex]?.isGap) previousIndex--;
-      let nextIndex = layerIndex + 1;
-      while (layers[nextIndex]?.isGap) nextIndex++;
-      const previousLayer = layers[previousIndex] ?? null;
-      const nextLayer = layers[nextIndex] ?? null;
-
-      allPossibleDepths = previousLayer ? allPossibleDepths.filter(d => d >= previousLayer.toDepth) : allPossibleDepths;
-      allPossibleDepths = nextLayer ? allPossibleDepths.filter(d => d <= nextLayer.fromDepth) : allPossibleDepths;
-
-      return getForToDepths ? allPossibleDepths.slice(1) : allPossibleDepths.slice(0, -1);
-    },
-    [tmpLithologiesFlat],
-  );
-
-  const isMutating =
-    isAddingLithology ||
-    isUpdatingLithology ||
-    isDeletingLithology ||
-    isAddingLithologicalDescription ||
-    isUpdatingLithologicalDescription ||
-    isDeletingLithologicalDescription ||
-    isAddingFaciesDescription ||
-    isUpdatingFaciesDescription ||
-    isDeletingFaciesDescription;
-
-  const initTmpLayers = useCallback(() => {
-    if (isMutating) return;
-    const initDepths = getLayerDepths(lithologies, lithologicalDescriptions, faciesDescriptions);
-
-    const tmpLithologies = getLayersWithGaps(
-      lithologies.map(layer => ({ item: layer, hasChanges: false })),
-      initDepths,
-      stratigraphyId,
-    );
-    setTmpLithologies(tmpLithologies);
-
-    const tmpLithologicalDescriptions = getLayersWithGaps(
-      lithologicalDescriptions.map(layer => ({ item: layer, hasChanges: false })),
-      initDepths,
-      stratigraphyId,
-    );
-    setTmpLithologicalDescriptions(tmpLithologicalDescriptions);
-
-    const tmpFaciesDescriptions = getLayersWithGaps(
-      faciesDescriptions.map(layer => ({ item: layer, hasChanges: false })),
-      initDepths,
-      stratigraphyId,
-    );
-    setTmpFaciesDescriptions(tmpFaciesDescriptions);
-    setDepths(initDepths);
-  }, [lithologies, lithologicalDescriptions, faciesDescriptions, stratigraphyId, isMutating]);
-
-  useEffect(() => {
-    if (isMutating) return;
-    if (tmpLithologies.length > 0 || tmpLithologicalDescriptions.length > 0 || tmpFaciesDescriptions.length > 0) {
-      const updatedDepths = getLayerDepths(
-        tmpLithologies.map(l => l.item as Lithology),
-        tmpLithologicalDescriptions.map(l => l.item as LithologicalDescription),
-        tmpFaciesDescriptions.map(l => l.item as FaciesDescription),
-      );
-      if (JSON.stringify(updatedDepths) !== JSON.stringify(depths)) {
-        setDepths(updatedDepths);
-      }
-
-      const newTmpLithologies = getLayersWithGaps(tmpLithologies, updatedDepths, stratigraphyId);
-      if (JSON.stringify(newTmpLithologies) !== JSON.stringify(tmpLithologies)) {
-        setTmpLithologies(newTmpLithologies);
-      }
-
-      const newTmpLithologicalDescriptions = getLayersWithGaps(
-        tmpLithologicalDescriptions,
-        updatedDepths,
-        stratigraphyId,
-      );
-      if (JSON.stringify(newTmpLithologicalDescriptions) !== JSON.stringify(tmpLithologicalDescriptions)) {
-        setTmpLithologicalDescriptions(newTmpLithologicalDescriptions);
-      }
-
-      const newTmpFaciesDescriptions = getLayersWithGaps(tmpFaciesDescriptions, updatedDepths, stratigraphyId);
-      if (JSON.stringify(newTmpFaciesDescriptions) !== JSON.stringify(tmpFaciesDescriptions)) {
-        setTmpFaciesDescriptions(newTmpFaciesDescriptions);
-      }
-    }
-  }, [tmpLithologies, tmpLithologicalDescriptions, tmpFaciesDescriptions, depths, stratigraphyId, isMutating]);
-
-  useEffect(() => {
-    initTmpLayers();
-  }, [initTmpLayers]);
-
-  const onReset = useCallback(async () => {
-    initTmpLayers();
-  }, [initTmpLayers]);
-
-  const deleteLithologies = useCallback(async () => {
-    for (const lithology of lithologies) {
-      if (!tmpLithologies.some(l => l.item.id === lithology.id)) {
-        await deleteLithology(lithology);
-      }
-    }
-  }, [deleteLithology, lithologies, tmpLithologies]);
-
-  const addAndUpdateLithologies = useCallback(async () => {
-    for (const lithology of tmpLithologies.filter(l => l.hasChanges).map(l => l.item as Lithology)) {
-      if (lithology.id === 0) {
-        await addLithology({ isGap: false, ...lithology, stratigraphyId });
-      } else {
-        await updateLithology({ isGap: false, ...lithology });
-      }
-    }
-  }, [addLithology, stratigraphyId, tmpLithologies, updateLithology]);
-
-  const deleteLithologicalDescriptions = useCallback(async () => {
-    for (const lithologicalDescription of lithologicalDescriptions) {
-      if (!tmpLithologicalDescriptions.some(l => l.item.id === lithologicalDescription.id)) {
-        await deleteLithologicalDescription(lithologicalDescription);
-      }
-    }
-  }, [deleteLithologicalDescription, lithologicalDescriptions, tmpLithologicalDescriptions]);
-
-  const addAndUpdateLithologicalDescriptions = useCallback(async () => {
-    for (const lithologicalDescription of tmpLithologicalDescriptions
-      .filter(l => l.hasChanges)
-      .map(l => l.item as LithologicalDescription)) {
-      if (lithologicalDescription.id === 0) {
-        await addLithologicalDescription({ isGap: false, ...lithologicalDescription, stratigraphyId });
-      } else {
-        await updateLithologicalDescription({ isGap: false, ...lithologicalDescription });
-      }
-    }
-  }, [addLithologicalDescription, stratigraphyId, tmpLithologicalDescriptions, updateLithologicalDescription]);
-
-  const deleteFaciesDescriptions = useCallback(async () => {
-    for (const faciesDescription of faciesDescriptions) {
-      if (!tmpFaciesDescriptions.some(l => l.item.id === faciesDescription.id)) {
-        await deleteFaciesDescription(faciesDescription);
-      }
-    }
-  }, [deleteFaciesDescription, faciesDescriptions, tmpFaciesDescriptions]);
-
-  const addAndUpdateFaciesDescriptions = useCallback(async () => {
-    for (const faciesDescription of tmpFaciesDescriptions
-      .filter(l => l.hasChanges)
-      .map(l => l.item as FaciesDescription)) {
-      if (faciesDescription.id === 0) {
-        await addFaciesDescription({ isGap: false, ...faciesDescription, stratigraphyId });
-      } else {
-        await updateFaciesDescription({ isGap: false, ...faciesDescription });
-      }
-    }
-  }, [addFaciesDescription, stratigraphyId, tmpFaciesDescriptions, updateFaciesDescription]);
-
-  const onSave = useCallback(async () => {
-    if (depths.some(c => c.hasFromDepthError || c.hasToDepthError)) {
-      showAlert(t(t("gapOrOverlayErrorCannotSave")), "error");
-      return false;
-    }
-    await Promise.all([
-      deleteLithologies(),
-      addAndUpdateLithologies(),
-      deleteLithologicalDescriptions(),
-      addAndUpdateLithologicalDescriptions(),
-      deleteFaciesDescriptions(),
-      addAndUpdateFaciesDescriptions(),
-    ]);
-    invalidateLithologyQueries();
-    invalidateLithologicalDescriptionQueries();
-    invalidateFaciesDescriptionQueries();
-    return true;
-  }, [
-    depths,
-    deleteLithologies,
-    addAndUpdateLithologies,
-    deleteLithologicalDescriptions,
-    addAndUpdateLithologicalDescriptions,
-    deleteFaciesDescriptions,
-    addAndUpdateFaciesDescriptions,
-    invalidateLithologyQueries,
-    invalidateLithologicalDescriptionQueries,
-    invalidateFaciesDescriptionQueries,
-    showAlert,
-    t,
-  ]);
-
-  useEffect(() => {
-    registerSaveHandler(onSave, "lithology");
-    registerResetHandler(onReset, "lithology");
-  }, [onReset, onSave, registerResetHandler, registerSaveHandler]);
-
-  const renderGapCell = (
-    index: number,
-    layer: BaseLayer,
-    keyPrefix: string,
-    onEdit: (index: number) => void,
-    GapComponent: typeof StratigraphyTableGap,
-  ) => {
-    return (
-      <GapComponent
-        key={`${keyPrefix}-${index}-${layer.fromDepth}-${layer.id}`}
-        dataCy={`${keyPrefix}-${layer.fromDepth}-${layer.id}`}
-        sx={{
-          height: `${computeCellHeight(layer.fromDepth, layer.toDepth, depths)}px`,
-        }}
-        index={index}
-        onClick={onEdit}
-      />
-    );
-  };
-
-  const renderActionCell = (
-    index: number,
-    layer: BaseLayer,
-    keyPrefix: string,
-    onEdit: (index: number) => void,
-    onDelete: (index: number) => void,
-    buildContent: (layer: BaseLayer) => ReactNode,
-  ) => (
-    <StratigraphyTableActionCell
-      key={`${keyPrefix}-${index}-${layer.fromDepth}-${layer.id}`}
-      dataCy={`${keyPrefix}-${layer.fromDepth}-${layer.toDepth}`}
-      sx={{
-        height: `${computeCellHeight(layer.fromDepth, layer.toDepth, depths)}px`,
-      }}
-      index={index}
-      onClick={onEdit}
-      onHoverClick={index => onDelete(index)}>
-      {buildContent(layer)}
-    </StratigraphyTableActionCell>
-  );
-
-  const isLithology = (layer: BaseLayer): layer is Lithology => {
-    return (layer as Lithology).isUnconsolidated !== undefined;
-  };
-
-  const renderTableCells = (
-    layers: BaseLayer[],
-    onEdit: (index: number) => void,
-    onDelete: (index: number) => void,
-    buildContent: (layer: BaseLayer) => ReactNode,
-    keyPrefix: string,
-    GapComponent: typeof StratigraphyTableGap,
-  ) => {
-    return layers.map((layer, index) => {
-      if (layer.isGap) {
-        const gapLayer: BaseLayer = { ...layer };
-        // only add isUnconsolidated if the layers are Lithology
-        if (isLithology(layer)) {
-          gapLayer.isUnconsolidated = (layers.at(index - 1) as Lithology)?.isUnconsolidated ?? true;
-        }
-        return renderGapCell(index, gapLayer, keyPrefix, onEdit, GapComponent);
-      }
-      return renderActionCell(index, layer, keyPrefix, onEdit, onDelete, buildContent);
-    });
-  };
-
-  return (
-    <>
-      <Stack gap={1.5}>
-        <Stack>
-          <StratigraphyTableHeader>
-            <StratigraphyTableHeaderCell sx={{ flex: "0 0 90px" }} label={t("depth")} />
-            <StratigraphyTableHeaderCell label={t("lithology")} />
-            <StratigraphyTableHeaderCell label={t("lithological_description")} />
-            <StratigraphyTableHeaderCell label={t("facies_description")} />
-          </StratigraphyTableHeader>
-          {depths?.length > 0 && (
-            <StratigraphyTableContent>
-              <StratigraphyTableColumn sx={{ flex: "0 0 90px" }}>
-                {/* TODO: Add FormInput for depths and update lithology if depth changes. https://github.com/swisstopo/swissgeol-boreholes-suite/issues/2392  Add overlap validation check */}
-                {depths.map(depth => (
-                  <StratigraphyTableCell
-                    key={`${depth.lithologyId}-depth-${depth.fromDepth}-${depth.toDepth}`}
-                    data-cy={`depth-${depth.fromDepth}-${depth.toDepth}`}
-                    sx={{ height: `${defaultRowHeight}px` }}>
-                    <Typography color={depth.hasFromDepthError ? "error" : "default"}>
-                      {`${formatNumberForDisplay(depth.fromDepth)} m MD`}
-                    </Typography>
-                    <Typography color={depth.hasToDepthError ? "error" : "default"}>
-                      {`${formatNumberForDisplay(depth.toDepth)} m MD`}
-                    </Typography>
-                  </StratigraphyTableCell>
-                ))}
-              </StratigraphyTableColumn>
-              <StratigraphyTableColumn>
-                {renderTableCells(
-                  tmpLithologiesFlat,
-                  handleEditLithology,
-                  handleDeleteLithology,
-                  layer => (
-                    <LithologyLabels lithology={layer as Lithology} />
-                  ),
-                  "lithology",
-                  StratigraphyTableGap,
-                )}
-              </StratigraphyTableColumn>
-              <StratigraphyTableColumn>
-                {renderTableCells(
-                  tmpLithologicalDescriptionsFlat,
-                  handleEditLithologicalDescription,
-                  handleDeleteLithologicalDescription,
-                  layer => (
-                    <Typography variant="body1" fontWeight={700}>
-                      {(layer as LithologicalDescription).description}
-                    </Typography>
-                  ),
-                  `lithologicalDescription`,
-                  StratigraphyTableDescriptionGap,
-                )}
-              </StratigraphyTableColumn>
-              <StratigraphyTableColumn>
-                {renderTableCells(
-                  tmpFaciesDescriptionsFlat,
-                  handleEditFaciesDescription,
-                  handleDeleteFaciesDescription,
-                  layer => (
-                    <FaciesDescriptionLabels description={layer as FaciesDescription} />
-                  ),
-                  `faciesDescription`,
-                  StratigraphyTableDescriptionGap,
-                )}
-              </StratigraphyTableColumn>
-            </StratigraphyTableContent>
-          )}
-        </Stack>
-        <AddRowButton
-          onClick={() => handleEditLithology(-1)}
-          dataCy="add-row-button"
-          buttonContent={<LayerAddButton />}
-        />
-      </Stack>
-      <LithologyModal lithology={selectedLithology} updateLithology={updateTmpLithology} />
-      <LithologicalDescriptionModal
-        description={selectedLithologicalDescription}
-        fromDepths={getDepthOptions(selectedLithologicalDescription, tmpLithologicalDescriptionsFlat)}
-        toDepths={getDepthOptions(selectedLithologicalDescription, tmpLithologicalDescriptionsFlat, true)}
-        updateLithologicalDescription={updateTmpLithologicalDescription}
-      />
-      <FaciesDescriptionModal
-        description={selectedFaciesDescription}
-        fromDepths={getDepthOptions(selectedFaciesDescription, tmpFaciesDescriptionsFlat)}
-        toDepths={getDepthOptions(selectedFaciesDescription, tmpFaciesDescriptionsFlat, true)}
-        updateFaciesDescription={updateTmpFaciesDescription}
-      />
-    </>
-  );
-};
