+++ src/client/src/pages/detail/form/stratigraphy/lithologyTable.tsx
@@ -1,14 +1,12 @@
-import { FC, MouseEvent, ReactNode, useCallback, useContext, useEffect, useRef, useState } from "react";
+import { FC, MouseEvent, ReactNode, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Stack, Typography } from "@mui/material";
 import Menu from "@mui/material/Menu";
 import MenuItem from "@mui/material/MenuItem";
 import { Trash2 } from "lucide-react";
-import { v4 as uuidv4 } from "uuid";
 import { BaseLayer } from "../../../../api/stratigraphy.ts";
 import { theme } from "../../../../AppTheme.ts";
 import { AddRowButton, StandaloneIconButton } from "../../../../components/buttons/buttons.tsx";
-import { SaveContext } from "../../saveContext.tsx";
 import { FaciesDescription } from "./faciesDescription.ts";
 import { LithologicalDescription } from "./lithologicalDescription.ts";
 import { Lithology } from "./lithology.ts";
@@ -18,15 +16,7 @@ import { FaciesDescriptionModal } from "./lithology/form/faciesDescriptionModal.
 import { LithologicalDescriptionModal } from "./lithology/form/lithologicalDescriptionModal.tsx";
 import { LithologyModal } from "./lithology/form/lithologyModal.tsx";
 import { LithologyLabels } from "./lithology/lithologyLabels.tsx";
-import {
-  createEmptyLithology,
-  defaultRowHeight,
-  DepthLayer,
-  flagErrors,
-  getInitialDepthLayers,
-  mergeAdjacentDepths,
-  removeDepthIdReferences,
-} from "./lithologyTableUtils.ts";
+import { defaultRowHeight } from "./lithologyTableUtils.ts";
 import {
   LayerAddButton,
   StratigraphyTableActionCell,
@@ -37,27 +27,37 @@ import {
   StratigraphyTableHeader,
   StratigraphyTableHeaderCell,
 } from "./stratigraphyTableComponents.tsx";
+import { LithologyTableState } from "./useLithologyTableState.ts";
+
+type LithologyTableColumn = "lithology" | "lithologicalDescription" | "faciesDescription";
+const allColumns: LithologyTableColumn[] = ["lithology", "lithologicalDescription", "faciesDescription"];
 
 interface LithologyTableProps {
-  lithologies: Lithology[];
-  lithologicalDescriptions: LithologicalDescription[];
-  faciesDescriptions: FaciesDescription[];
-  stratigraphyId: number;
+  state: LithologyTableState;
+  shownColumns?: LithologyTableColumn[];
 }
 
-export const LithologyTable: FC<LithologyTableProps> = ({
-  lithologies,
-  lithologicalDescriptions,
-  faciesDescriptions,
-  stratigraphyId,
-}) => {
+export const LithologyTable: FC<LithologyTableProps> = ({ state, shownColumns = allColumns }) => {
   const { t } = useTranslation();
-  const { markAsChanged } = useContext(SaveContext);
+  const {
+    depths,
+    tmpLithologies,
+    tmpLithologicalDescriptions,
+    tmpFaciesDescriptions,
+    stratigraphyId,
+    updateDepthBoundaries,
+    handleAddDepthLayer,
+    handleDeleteDepthLayer,
+    handleDeleteDescription,
+    updateTmpLithology,
+    updateTmpLithologicalDescription,
+    updateTmpFaciesDescription,
+  } = state;
 
-  const [depths, setDepths] = useState<DepthLayer[]>([]);
-  const [tmpLithologies, setTmpLithologies] = useState<Lithology[]>([]);
-  const [tmpLithologicalDescriptions, setTmpLithologicalDescriptions] = useState<LithologicalDescription[]>([]);
-  const [tmpFaciesDescriptions, setTmpFaciesDescriptions] = useState<FaciesDescription[]>([]);
+  const shown = useMemo(() => new Set(shownColumns), [shownColumns]);
+  const showLithology = shown.has("lithology");
+  const showLithologicalDescription = shown.has("lithologicalDescription");
+  const showFaciesDescription = shown.has("faciesDescription");
 
   const [selectedLithology, setSelectedLithology] = useState<Lithology>();
   const [selectedLithologicalDescription, setSelectedLithologicalDescription] = useState<LithologicalDescription>();
@@ -72,348 +72,41 @@ export const LithologyTable: FC<LithologyTableProps> = ({
     layer.fromDepth === previewDeleteDepth.fromDepth &&
     layer.toDepth === previewDeleteDepth.toDepth;
 
-  const baselineRef = useRef({
-    lithologies: "[]",
-    lithologicalDescriptions: "[]",
-    faciesDescriptions: "[]",
-  });
+  const handleLithologyUpdate = (updated: Lithology, hasChanges: boolean) => {
+    updateTmpLithology(updated, hasChanges);
+    setSelectedLithology(undefined);
+  };
 
-  useEffect(() => {
-    const { cleanDepths, cleanLithologies, cleanLithologicalDescriptions, cleanFaciesDescriptions } =
-      getInitialDepthLayers(lithologies, lithologicalDescriptions, faciesDescriptions, stratigraphyId);
-    baselineRef.current = {
-      lithologies: JSON.stringify(cleanLithologies),
-      lithologicalDescriptions: JSON.stringify(cleanLithologicalDescriptions),
-      faciesDescriptions: JSON.stringify(cleanFaciesDescriptions),
-    };
-    setDepths(cleanDepths);
-    setTmpLithologies(cleanLithologies);
-    setTmpLithologicalDescriptions(cleanLithologicalDescriptions);
-    setTmpFaciesDescriptions(cleanFaciesDescriptions);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  const handleLithologicalDescriptionUpdate = (updated: LithologicalDescription, hasChanges: boolean) => {
+    updateTmpLithologicalDescription(updated, hasChanges);
+    setSelectedLithologicalDescription(undefined);
+  };
 
-  const commitChanges = useCallback(
-    (
-      newDepths: DepthLayer[],
-      newLithologies: Lithology[],
-      newLithologicalDescriptions: LithologicalDescription[],
-      newFaciesDescriptions: FaciesDescription[],
-    ) => {
-      const newLithologiesJson = JSON.stringify(newLithologies);
-      const newLithologicalDescriptionsJson = JSON.stringify(newLithologicalDescriptions);
-      const newFaciesDescriptionsJson = JSON.stringify(newFaciesDescriptions);
+  const handleFaciesDescriptionUpdate = (updated: FaciesDescription, hasChanges: boolean) => {
+    updateTmpFaciesDescription(updated, hasChanges);
+    setSelectedFaciesDescription(undefined);
+  };
 
-      const depthsChanged = JSON.stringify(newDepths) !== JSON.stringify(depths);
-      const lithologiesChanged = newLithologiesJson !== JSON.stringify(tmpLithologies);
-      const lithologicalDescriptionsChanged =
-        newLithologicalDescriptionsJson !== JSON.stringify(tmpLithologicalDescriptions);
-      const faciesDescriptionsChanged = newFaciesDescriptionsJson !== JSON.stringify(tmpFaciesDescriptions);
+  const handleAddLithologicalDescriptionInGap = (depthId: string, fromDepth: number, toDepth: number) => {
+    setSelectedLithologicalDescription({
+      id: 0,
+      stratigraphyId,
+      fromDepth,
+      toDepth,
+      depthIds: [depthId],
+    });
+  };
 
-      if (!depthsChanged && !lithologiesChanged && !lithologicalDescriptionsChanged && !faciesDescriptionsChanged) {
-        return;
-      }
-
-      if (depthsChanged) setDepths(newDepths);
-      if (lithologiesChanged) setTmpLithologies(newLithologies);
-      if (lithologicalDescriptionsChanged) setTmpLithologicalDescriptions(newLithologicalDescriptions);
-      if (faciesDescriptionsChanged) setTmpFaciesDescriptions(newFaciesDescriptions);
-
-      const matchesBaseline =
-        newLithologiesJson === baselineRef.current.lithologies &&
-        newLithologicalDescriptionsJson === baselineRef.current.lithologicalDescriptions &&
-        newFaciesDescriptionsJson === baselineRef.current.faciesDescriptions;
-      markAsChanged(!matchesBaseline);
-    },
-    [depths, tmpLithologies, tmpLithologicalDescriptions, tmpFaciesDescriptions, markAsChanged],
-  );
-
-  const updateDepthBoundaries = useCallback(
-    (depthId: string, side: "from" | "to", newValue: number) => {
-      const primaryIndex = depths.findIndex(d => d.id === depthId);
-      if (primaryIndex < 0) return;
-      const primary = depths[primaryIndex];
-      const oldValue = side === "from" ? primary.fromDepth : primary.toDepth;
-      if (oldValue === newValue) return;
-
-      const sideUpdates = new Map<string, "from" | "to">();
-      sideUpdates.set(depthId, side);
-
-      // Bring the immediately adjacent depth layer along on the opposite side, since it shares the boundary.
-      if (side === "to") {
-        const next = depths[primaryIndex + 1];
-        if (next && next.fromDepth === oldValue) sideUpdates.set(next.id, "from");
-      } else {
-        const prev = depths[primaryIndex - 1];
-        if (prev && prev.toDepth === oldValue) sideUpdates.set(prev.id, "to");
-      }
-
-      // Update depth layers
-      const newDepths = depths.map(d => {
-        const updSide = sideUpdates.get(d.id);
-        if (!updSide) return d;
-        return updSide === "from" ? { ...d, fromDepth: newValue } : { ...d, toDepth: newValue };
-      });
-
-      // Update data layers
-      const propagate = <T extends BaseLayer>(items: T[]): T[] =>
-        items.map(item => {
-          if (!item.depthIds || item.depthIds.length === 0) return item;
-          let newFromDepth = item.fromDepth;
-          let newToDepth = item.toDepth;
-          let changed = false;
-          for (const id of item.depthIds) {
-            const updSide = sideUpdates.get(id);
-            if (!updSide) continue;
-            if (updSide === "from" && item.fromDepth === oldValue) {
-              newFromDepth = newValue;
-              changed = true;
-            } else if (updSide === "to" && item.toDepth === oldValue) {
-              newToDepth = newValue;
-              changed = true;
-            }
-          }
-          if (!changed) return item;
-          return { ...item, fromDepth: newFromDepth, toDepth: newToDepth };
-        });
-
-      const newLithologies = propagate(tmpLithologies);
-      const newLithologicalDescriptions = propagate(tmpLithologicalDescriptions);
-      const newFaciesDescriptions = propagate(tmpFaciesDescriptions);
-
-      commitChanges(
-        flagErrors(newDepths, newLithologies),
-        newLithologies,
-        newLithologicalDescriptions,
-        newFaciesDescriptions,
-      );
-    },
-    [depths, tmpLithologies, tmpLithologicalDescriptions, tmpFaciesDescriptions, commitChanges],
-  );
-
-  const handleAddDepthLayer = useCallback(() => {
-    const lastDepth = depths[depths.length - 1];
-    const lastLithology = tmpLithologies[tmpLithologies.length - 1];
-    const newBoundary = lastDepth?.toDepth ?? 0;
-    const newDepthLayer: DepthLayer = {
-      id: uuidv4(),
-      fromDepth: newBoundary,
-      toDepth: newBoundary,
-    };
-    const newLithology: Lithology = {
-      ...createEmptyLithology(newBoundary, newBoundary, stratigraphyId, lastLithology?.isUnconsolidated ?? true),
-      depthIds: [newDepthLayer.id],
-    };
-    const newDepths = [...depths, newDepthLayer];
-    const newLithologies = [...tmpLithologies, newLithology];
-    commitChanges(
-      flagErrors(newDepths, newLithologies),
-      newLithologies,
-      tmpLithologicalDescriptions,
-      tmpFaciesDescriptions,
-    );
-  }, [depths, tmpLithologies, tmpLithologicalDescriptions, tmpFaciesDescriptions, stratigraphyId, commitChanges]);
-
-  // Delete a description and merge any depth layers it referenced whose exact range is no
-  // longer matched by any remaining item. The previous adjacent depth layer absorbs the
-  // candidate (its toDepth is extended). Depth-layer ids on the survivors stay put; items
-  // just have the absorbed-layer ids dropped from their depthIds.
-  const handleDeleteDescription = useCallback(
-    (kind: "lithological" | "facies", index: number) => {
-      const list = kind === "lithological" ? tmpLithologicalDescriptions : tmpFaciesDescriptions;
-      if (index < 0 || index >= list.length) return;
-      const deletedItem = list[index];
-      const trimmed = [...list.slice(0, index), ...list.slice(index + 1)];
-      const newLithologicalDescriptions = kind === "lithological" ? trimmed : tmpLithologicalDescriptions;
-      const newFaciesDescriptions = kind === "facies" ? trimmed : tmpFaciesDescriptions;
-
-      // Exact (fromDepth-toDepth) ranges still claimed by some remaining item.
-      const ownedRanges = new Set<string>();
-      for (const items of [tmpLithologies, newLithologicalDescriptions, newFaciesDescriptions] as BaseLayer[][]) {
-        for (const item of items) {
-          ownedRanges.add(`${item.fromDepth}-${item.toDepth}`);
-        }
-      }
-
-      // Merge candidates: depth layers the deleted item referenced whose range is no longer
-      // owned (no remaining item has matching fromDepth+toDepth).
-      const candidates = new Set<string>();
-      for (const id of deletedItem.depthIds ?? []) {
-        const depth = depths.find(d => d.id === id);
-        if (depth && !ownedRanges.has(`${depth.fromDepth}-${depth.toDepth}`)) {
-          candidates.add(id);
-        }
-      }
-
-      const { depths: mergedDepths, mergedIds } = mergeAdjacentDepths(depths, candidates);
-
-      const updatedLithologies = removeDepthIdReferences(tmpLithologies, mergedIds);
-      const updatedLithologicalDescriptions = removeDepthIdReferences(newLithologicalDescriptions, mergedIds);
-      const updatedFaciesDescriptions = removeDepthIdReferences(newFaciesDescriptions, mergedIds);
-
-      commitChanges(
-        flagErrors(mergedDepths, updatedLithologies),
-        updatedLithologies,
-        updatedLithologicalDescriptions,
-        updatedFaciesDescriptions,
-      );
-    },
-    [depths, tmpLithologies, tmpLithologicalDescriptions, tmpFaciesDescriptions, commitChanges],
-  );
-
-  const mergeModalUpdate = useCallback(
-    <T extends BaseLayer>(tmpItems: T[], updated: T): T[] => {
-      const depthIdsKey = JSON.stringify(updated.depthIds ?? []);
-      const idx = tmpItems.findIndex(
-        item =>
-          item.fromDepth === updated.fromDepth &&
-          item.toDepth === updated.toDepth &&
-          JSON.stringify(item.depthIds ?? []) === depthIdsKey,
-      );
-      if (idx >= 0) return tmpItems.map((item, i) => (i === idx ? updated : item));
-
-      const depthOrder = new Map<string, number>();
-      depths.forEach((d, i) => depthOrder.set(d.id, i));
-      const firstDepthIdx = (item: T) =>
-        Math.min(...(item.depthIds?.map(id => depthOrder.get(id) ?? Infinity) ?? [Infinity]));
-      return [...tmpItems, updated].sort((a, b) => firstDepthIdx(a) - firstDepthIdx(b));
-    },
-    [depths],
-  );
-
-  const updateTmpLithology = useCallback(
-    (updated: Lithology, hasChanges: boolean) => {
-      if (hasChanges) {
-        const newLithologies = mergeModalUpdate(tmpLithologies, updated);
-        commitChanges(depths, newLithologies, tmpLithologicalDescriptions, tmpFaciesDescriptions);
-      }
-      setSelectedLithology(undefined);
-    },
-    [mergeModalUpdate, tmpLithologies, depths, tmpLithologicalDescriptions, tmpFaciesDescriptions, commitChanges],
-  );
-
-  const updateTmpLithologicalDescription = useCallback(
-    (updated: LithologicalDescription, hasChanges: boolean) => {
-      if (hasChanges) {
-        const newDescs = mergeModalUpdate(tmpLithologicalDescriptions, updated);
-        commitChanges(depths, tmpLithologies, newDescs, tmpFaciesDescriptions);
-      }
-      setSelectedLithologicalDescription(undefined);
-    },
-    [mergeModalUpdate, tmpLithologicalDescriptions, depths, tmpLithologies, tmpFaciesDescriptions, commitChanges],
-  );
-
-  const updateTmpFaciesDescription = useCallback(
-    (updated: FaciesDescription, hasChanges: boolean) => {
-      if (hasChanges) {
-        const newDescs = mergeModalUpdate(tmpFaciesDescriptions, updated);
-        commitChanges(depths, tmpLithologies, tmpLithologicalDescriptions, newDescs);
-      }
-      setSelectedFaciesDescription(undefined);
-    },
-    [mergeModalUpdate, tmpFaciesDescriptions, depths, tmpLithologies, tmpLithologicalDescriptions, commitChanges],
-  );
-
-  const handleAddLithologicalDescriptionInGap = useCallback(
-    (depthId: string, fromDepth: number, toDepth: number) => {
-      setSelectedLithologicalDescription({
-        id: 0,
-        stratigraphyId,
-        fromDepth,
-        toDepth,
-        depthIds: [depthId],
-      });
-    },
-    [stratigraphyId],
-  );
-
-  const handleAddFaciesDescriptionInGap = useCallback(
-    (depthId: string, fromDepth: number, toDepth: number) => {
-      setSelectedFaciesDescription({
-        id: 0,
-        stratigraphyId,
-        fromDepth,
-        toDepth,
-        faciesId: null,
-        depthIds: [depthId],
-      });
-    },
-    [stratigraphyId],
-  );
-
-  const handleDeleteDepthLayer = useCallback(
-    (depthId: string, action: "extendLower" | "extendUpper" | "reduceBoreholeEnd") => {
-      if (depths.length === 1) {
-        commitChanges([], [], [], []);
-        return;
-      }
-
-      const index = depths.findIndex(d => d.id === depthId);
-      if (index < 0) return;
-      const depthLayerToDelete = depths[index];
-      const depthLayerToUpdate =
-        action === "extendUpper" ? depths[index - 1] : action === "extendLower" ? depths[index + 1] : null;
-
-      const updatedDepthLayers = depths.flatMap((d, i) => {
-        if (i === index) return [];
-        if (action === "extendLower" && i === index + 1) return [{ ...d, fromDepth: depthLayerToDelete.fromDepth }];
-        if (action === "extendUpper" && i === index - 1) return [{ ...d, toDepth: depthLayerToDelete.toDepth }];
-        return [{ ...d }];
-      });
-
-      // Remove layers that completely match deleted depth layer
-      const matchesDeletedDepths = <T extends BaseLayer>(item: T) =>
-        item.depthIds?.includes(depthLayerToDelete.id) &&
-        item.fromDepth === depthLayerToDelete.fromDepth &&
-        item.toDepth === depthLayerToDelete.toDepth;
-
-      const updateItem = <T extends BaseLayer>(item: T): T => {
-        // Remove deleted depth layer id reference
-        const newDepthIds = item.depthIds?.filter(id => id !== depthLayerToDelete.id);
-
-        // Adjust items linked to the depth layer to update
-        if (depthLayerToUpdate && item.depthIds?.includes(depthLayerToUpdate.id)) {
-          if (action === "extendUpper" && item.toDepth === depthLayerToUpdate.toDepth) {
-            return { ...item, toDepth: depthLayerToDelete.toDepth, depthIds: newDepthIds };
-          }
-          if (action === "extendLower" && item.fromDepth === depthLayerToUpdate.fromDepth) {
-            return { ...item, fromDepth: depthLayerToDelete.fromDepth, depthIds: newDepthIds };
-          }
-        }
-
-        // Shift items linked to deleted depth layer whose boundary sat on the layer's disappearing edge.
-        if (item.depthIds?.includes(depthLayerToDelete.id)) {
-          if (action === "extendUpper" && item.fromDepth === depthLayerToDelete.fromDepth) {
-            return { ...item, fromDepth: depthLayerToDelete.toDepth, depthIds: newDepthIds };
-          }
-          if (
-            (action === "extendLower" || action === "reduceBoreholeEnd") &&
-            item.toDepth === depthLayerToDelete.toDepth
-          ) {
-            return { ...item, toDepth: depthLayerToDelete.fromDepth, depthIds: newDepthIds };
-          }
-        }
-
-        // Keep items unrelated to the deleted or updated depth layers unchanged
-        if (newDepthIds?.length === item.depthIds?.length) return item;
-        return { ...item, depthIds: newDepthIds };
-      };
-
-      const newLithologies = tmpLithologies.filter(l => !matchesDeletedDepths(l)).map(updateItem);
-      const newLithologicalDescriptions = tmpLithologicalDescriptions
-        .filter(d => !matchesDeletedDepths(d))
-        .map(updateItem);
-      const newFaciesDescriptions = tmpFaciesDescriptions.filter(d => !matchesDeletedDepths(d)).map(updateItem);
-
-      commitChanges(
-        flagErrors(updatedDepthLayers, newLithologies),
-        newLithologies,
-        newLithologicalDescriptions,
-        newFaciesDescriptions,
-      );
-    },
-    [depths, tmpLithologies, tmpLithologicalDescriptions, tmpFaciesDescriptions, commitChanges],
-  );
+  const handleAddFaciesDescriptionInGap = (depthId: string, fromDepth: number, toDepth: number) => {
+    setSelectedFaciesDescription({
+      id: 0,
+      stratigraphyId,
+      fromDepth,
+      toDepth,
+      faciesId: null,
+      depthIds: [depthId],
+    });
+  };
 
   const renderGapCell = (
     index: number,
@@ -498,9 +191,9 @@ export const LithologyTable: FC<LithologyTableProps> = ({
       <Stack>
         <StratigraphyTableHeader>
           <StratigraphyTableHeaderCell sx={{ flex: "0 0 120px" }} label={t("depth")} />
-          <StratigraphyTableHeaderCell label={t("lithology")} />
-          <StratigraphyTableHeaderCell label={t("lithological_description")} />
-          <StratigraphyTableHeaderCell label={t("facies_description")} />
+          {showLithology && <StratigraphyTableHeaderCell label={t("lithology")} />}
+          {showLithologicalDescription && <StratigraphyTableHeaderCell label={t("lithological_description")} />}
+          {showFaciesDescription && <StratigraphyTableHeaderCell label={t("facies_description")} />}
         </StratigraphyTableHeader>
         {depths?.length > 0 && (
           <StratigraphyTableContent>
@@ -642,54 +335,60 @@ export const LithologyTable: FC<LithologyTableProps> = ({
                 })()}
               </Menu>
             </StratigraphyTableColumn>
-            <StratigraphyTableColumn>
-              {renderTableCells(
-                "lithology",
-                tmpLithologies,
-                layer => (
-                  <LithologyLabels lithology={layer as Lithology} />
-                ),
-                index => setSelectedLithology(tmpLithologies[index]),
-              )}
-            </StratigraphyTableColumn>
-            <StratigraphyTableColumn>
-              {renderTableCells(
-                `lithologicalDescription`,
-                tmpLithologicalDescriptions,
-                layer => (
-                  <Typography variant="body1" fontWeight={700}>
-                    {(layer as LithologicalDescription).description}
-                  </Typography>
-                ),
-                index => setSelectedLithologicalDescription(tmpLithologicalDescriptions[index]),
-                index => handleDeleteDescription("lithological", index),
-                handleAddLithologicalDescriptionInGap,
-              )}
-            </StratigraphyTableColumn>
-            <StratigraphyTableColumn>
-              {renderTableCells(
-                `faciesDescription`,
-                tmpFaciesDescriptions,
-                layer => (
-                  <FaciesDescriptionLabels description={layer as FaciesDescription} />
-                ),
-                index => setSelectedFaciesDescription(tmpFaciesDescriptions[index]),
-                index => handleDeleteDescription("facies", index),
-                handleAddFaciesDescriptionInGap,
-              )}
-            </StratigraphyTableColumn>
+            {showLithology && (
+              <StratigraphyTableColumn>
+                {renderTableCells(
+                  "lithology",
+                  tmpLithologies,
+                  layer => (
+                    <LithologyLabels lithology={layer as Lithology} />
+                  ),
+                  index => setSelectedLithology(tmpLithologies[index]),
+                )}
+              </StratigraphyTableColumn>
+            )}
+            {showLithologicalDescription && (
+              <StratigraphyTableColumn>
+                {renderTableCells(
+                  `lithologicalDescription`,
+                  tmpLithologicalDescriptions,
+                  layer => (
+                    <Typography variant="body1" fontWeight={700}>
+                      {(layer as LithologicalDescription).description}
+                    </Typography>
+                  ),
+                  index => setSelectedLithologicalDescription(tmpLithologicalDescriptions[index]),
+                  index => handleDeleteDescription("lithological", index),
+                  handleAddLithologicalDescriptionInGap,
+                )}
+              </StratigraphyTableColumn>
+            )}
+            {showFaciesDescription && (
+              <StratigraphyTableColumn>
+                {renderTableCells(
+                  `faciesDescription`,
+                  tmpFaciesDescriptions,
+                  layer => (
+                    <FaciesDescriptionLabels description={layer as FaciesDescription} />
+                  ),
+                  index => setSelectedFaciesDescription(tmpFaciesDescriptions[index]),
+                  index => handleDeleteDescription("facies", index),
+                  handleAddFaciesDescriptionInGap,
+                )}
+              </StratigraphyTableColumn>
+            )}
           </StratigraphyTableContent>
         )}
       </Stack>
       <AddRowButton onClick={handleAddDepthLayer} dataCy="add-row-button" buttonContent={<LayerAddButton />} />
-      <LithologyModal lithology={selectedLithology} updateLithology={updateTmpLithology} />
+      <LithologyModal lithology={selectedLithology} updateLithology={handleLithologyUpdate} />
       <LithologicalDescriptionModal
         description={selectedLithologicalDescription}
-        updateLithologicalDescription={updateTmpLithologicalDescription}
+        updateLithologicalDescription={handleLithologicalDescriptionUpdate}
       />
       <FaciesDescriptionModal
         description={selectedFaciesDescription}
-        updateFaciesDescription={updateTmpFaciesDescription}
+        updateFaciesDescription={handleFaciesDescriptionUpdate}
       />
     </Stack>
   );
