+++ src/client/src/pages/detail/form/stratigraphy/stratigraphyUtils.ts
@@ -1,6 +1,4 @@
 import { BaseLayer } from "../../../../api/stratigraphy.ts";
-import { FaciesDescription } from "./faciesDescription.ts";
-import { LithologicalDescription } from "./lithologicalDescription.ts";
 import { LayerDepth, Lithology } from "./lithology.ts";
 
 export const defaultRowHeight = 240;
@@ -10,184 +8,6 @@ export const computeCellHeight = (fromDepth: number, toDepth: number, depths: La
   const endIndex = depths.findIndex(l => l.toDepth === toDepth);
   if (startIndex === -1 || endIndex === -1) return defaultRowHeight;
   return (endIndex - startIndex + 1) * defaultRowHeight;
-};
-
-const createGapLayerDepth = (fromDepth: number, toDepth: number): LayerDepth => ({
-  fromDepth,
-  toDepth,
-  lithologyId: 0,
-  hasFromDepthError: true,
-  hasToDepthError: true,
-});
-
-const checkDescriptionDepthExtendingCurrentLayers = (descriptions: BaseLayer[], layerDepths: LayerDepth[]) => {
-  const firstLayer = layerDepths.at(0)!;
-  const lastLayer = layerDepths.at(-1)!;
-
-  const firstDescription = descriptions.at(0)!;
-  const lastDescription = descriptions.at(-1)!;
-
-  if (firstDescription.fromDepth < firstLayer.fromDepth) {
-    layerDepths.push(createGapLayerDepth(firstDescription.fromDepth, firstLayer.fromDepth));
-  }
-
-  if (lastLayer.toDepth < lastDescription.toDepth) {
-    layerDepths.push(createGapLayerDepth(lastLayer.toDepth, lastDescription.toDepth));
-  }
-};
-
-const checkLayerForMultipleDescriptionDepthsFullyWithin = (descriptions: BaseLayer[], layerDepths: LayerDepth[]) => {
-  for (let i = 0; i < layerDepths.length; i++) {
-    const layer = layerDepths[i];
-    if (layer.lithologyId === 0) {
-      // Find descriptions fully within this layer
-      const matchingDescriptions = descriptions.filter(
-        d => d.fromDepth >= layer.fromDepth && d.toDepth <= layer.toDepth,
-      );
-      if (matchingDescriptions.length > 1) {
-        // Remove the original layer
-        layerDepths.splice(i, 1);
-        // Insert new layers for each description
-        const newLayers = matchingDescriptions.map(d => createGapLayerDepth(d.fromDepth, d.toDepth));
-        layerDepths.splice(i, 0, ...newLayers);
-        i += newLayers.length - 1;
-      }
-    }
-  }
-};
-
-const checkLayerForDescriptionDepthOverlapping = (descriptions: BaseLayer[], layerDepths: LayerDepth[]) => {
-  for (let i = 0; i < layerDepths.length; i++) {
-    const layer = layerDepths[i];
-    if (layer.lithologyId === 0) {
-      // Find descriptions that overlap this layer, but are not exact matches of fromDepth and toDepth with any other layers
-      const overlappingDescriptions = descriptions.filter(
-        d =>
-          !(layerDepths.some(l => d.fromDepth === l.fromDepth) && layerDepths.some(l => d.toDepth === l.toDepth)) &&
-          d.fromDepth < layer.toDepth &&
-          d.toDepth > layer.fromDepth,
-      );
-      if (overlappingDescriptions.length > 0) {
-        // For each overlapping description, split the gap as needed
-        const splits: LayerDepth[] = [];
-        let start = layer.fromDepth;
-        for (const desc of overlappingDescriptions) {
-          if (desc.fromDepth > start) {
-            splits.push(createGapLayerDepth(start, desc.fromDepth));
-          }
-          splits.push(createGapLayerDepth(Math.max(start, desc.fromDepth), Math.min(layer.toDepth, desc.toDepth)));
-          start = Math.min(layer.toDepth, desc.toDepth);
-        }
-        if (start < layer.toDepth) {
-          splits.push(createGapLayerDepth(start, layer.toDepth));
-        }
-        // Remove the original gap and insert the splits
-        layerDepths.splice(i, 1, ...splits);
-        i += splits.length - 1;
-      }
-    }
-  }
-};
-
-const checkDescriptionDepthLayers = (descriptions: BaseLayer[], layerDepths: LayerDepth[]) => {
-  if (descriptions.length === 0) return;
-
-  // If there are no current layers, add all descriptions as gap layers
-  if (layerDepths.length === 0) {
-    for (const description of descriptions) {
-      layerDepths.push(createGapLayerDepth(description.fromDepth, description.toDepth));
-    }
-    return;
-  }
-
-  // Add gap layers if descriptions extend beyond current layers
-  checkDescriptionDepthExtendingCurrentLayers(descriptions, layerDepths);
-
-  // Split gap layers if multiple descriptions fit inside a layer
-  checkLayerForMultipleDescriptionDepthsFullyWithin(descriptions, layerDepths);
-
-  // Split gap layers if a description overlaps the layer
-  checkLayerForDescriptionDepthOverlapping(descriptions, layerDepths);
-};
-
-const checkLayerDepths = (depths: LayerDepth[]) => {
-  const checkedLayerDepths: LayerDepth[] = [];
-  for (let i = 0; i < depths.length; i++) {
-    const previous = depths[i - 1];
-    const current = depths[i];
-    const next = depths[i + 1];
-
-    const hasFromDepthOverlap = (previous ? current.fromDepth < previous.toDepth : false) || current.hasFromDepthError;
-    const hasToDepthOverlap = (next ? current.toDepth > next.fromDepth : false) || current.hasToDepthError;
-
-    checkedLayerDepths.push({
-      ...current,
-      hasFromDepthError: hasFromDepthOverlap,
-      hasToDepthError: hasToDepthOverlap,
-    });
-
-    // Add gap layer if there's a gap between current and next. This can happen if the lithologies and
-    // lithologicalDescriptions are empty but the faciesDescriptions contains layers with gaps in between.
-    if (i < depths.length - 1) {
-      if (current.toDepth < next.fromDepth) {
-        checkedLayerDepths.push({
-          fromDepth: current.toDepth,
-          toDepth: next.fromDepth,
-          lithologyId: 0,
-          hasFromDepthError: true,
-          hasToDepthError: true,
-        });
-      }
-    }
-  }
-  return checkedLayerDepths;
-};
-
-export const getLayerDepths = (
-  lithologies: Lithology[],
-  lithologicalDescriptions: LithologicalDescription[],
-  faciesDescriptions: FaciesDescription[],
-) => {
-  const lithologiesWithoutGaps = lithologies.filter(l => !l.isGap).sort((a, b) => a.fromDepth - b.fromDepth);
-  const layerDepths: LayerDepth[] = [];
-  for (let i = 0; i < lithologiesWithoutGaps.length; i++) {
-    const current = lithologiesWithoutGaps[i];
-    // Add gap layer for gap between previous and current lithology
-    if (i > 0) {
-      const prev = lithologiesWithoutGaps[i - 1];
-      if (current.fromDepth > prev.toDepth) {
-        layerDepths.push({
-          fromDepth: prev.toDepth,
-          toDepth: current.fromDepth,
-          lithologyId: 0,
-          hasFromDepthError: true,
-          hasToDepthError: true,
-        });
-      }
-    }
-    layerDepths.push({
-      fromDepth: current.fromDepth,
-      toDepth: current.toDepth,
-      lithologyId: current.id,
-      hasFromDepthError: false,
-      hasToDepthError: false,
-    });
-  }
-  layerDepths.sort((a, b) => a.fromDepth - b.fromDepth);
-
-  checkDescriptionDepthLayers(
-    lithologicalDescriptions.filter(l => !l.isGap),
-    layerDepths,
-  );
-  layerDepths.sort((a, b) => a.fromDepth - b.fromDepth);
-
-  checkDescriptionDepthLayers(
-    faciesDescriptions.filter(l => !l.isGap),
-    layerDepths,
-  );
-  layerDepths.sort((a, b) => a.fromDepth - b.fromDepth);
-
-  return checkLayerDepths(layerDepths);
 };
 
 const isLithology = (layer: BaseLayer): layer is Lithology => {
@@ -309,7 +129,7 @@ export const getLayersWithGaps = (
   return resultLayers;
 };
 
-export interface BaseLayerChangeTracker {
+interface BaseLayerChangeTracker {
   item: BaseLayer;
   hasChanges: boolean;
 }
